+++ packages/cf-deploy-config-writer/src/index.ts
@@ -1,4 +1,4 @@
 export * from './mta-config';
 export * from './cf-writer';
 export { DefaultMTADestination } from './constants';
-export { CFBaseConfig, CFAppConfig, RouterModuleType, ApiHubConfig, ApiHubType } from './types';
+export { CFBaseConfig, CFAppConfig, CAPConfig, RouterModuleType, ApiHubConfig, ApiHubType } from './types';